refactor: dynamically import client components
